Rewrite the entire bot
This was a long time coming and a lot of stuff has changed.

At a high level, there are four major changes:
1. Total restructuring so it's way more clean and not an Italian
   restaurant's worth of spaghetti
2. Now using the current Discord API as well as just a general
   upgrading of the dependencies
3. DB schema changes. The old DB relied on storing Discord message IDs
   in order to clean up old messages. The new one just scans through the
   channel. This is WAY easier to handle because it means the bot is not
   trying to store Discord state AND reddit state.
4. Docker! The old deploy was a little silly. But now with the power of
   Docker, builds + deploys are dead simple. Huzzah.

Here's each of the changes in a bit more detail.

** Restructure **

There is now a brand new structure to the project.
`bot.py` is used just to invoke the actual bot.
All the actual bot code is in `src/`.

The things that weren't Discord-specific changes are basically:

1. Better use of modules
2. Better use of classes + encapsulation instead of spaghetti
3. DB is now a singleton instance

The lack of spaghetti is the big one. This is so much easier to follow
now.

** Discord API **

The old version of the bot was using a SUPER old version of Discord's
official python module. This now is using the current version. Yay.

That means a few things.

For one, separation of concerns into Cogs. What was previously just a massive
file of all the client logic in it, we now have four Cogs.

1. Cleanup -- Responsible for syncing the state of the channels with
   actions done directly on reddit rather than through the bot.
2. Commands -- Responsible for the actual commands that the bot exposes
   for mods to use on the subreddit. Namely approving/removing posts and
   comments with/without reasons.
3. PostsStream -- Responsible for polling the mod queue and showing any
   posts that have not been actioned on by a mod.
4. ReportsStream -- Responsible for polling the mod reports queue and
   showing any new reports that have been filed along with reasons.

** DB changes **

This one is relatively straightforward but basically now there are
really only two columns in each of the two tables: submission ID and
resolved.

Previously, the DB stored Discord message snowflakes so that
it could issue requests without needing to actually scan through the
channel message history to find old messages to clean up. But that just
over-complicated things for an optimization that really wasn't worth it.
The new way is much easier.

** Docker **

Yay Docker. I was late to the game with Docker so now I'm probably
over-using it but it makes development and deploys super easy so why
not.

We now have a single container for building + running the bot and a
mounted volume for storing the sqlite DB.

This has not only made development easy it will also make deploys an
absolute breeze.